Text.Numeral.Grammar

Inflection

data Inflection

Constructors

Inflection

Fields

iCase :: Maybe Case

iGender :: Maybe Gender

iNumber :: Maybe Number

defaultInflection :: Inflection

Grammatical categories

data Case

Constructors

Abessive
In linguistics, abessive (abbreviated abe or abess), caritive and privative (abbreviated priv) are names for a grammatical case expressing the lack or absence of the marked noun.

Ablative
The ablative case (abbreviated abl) indicates movement from something, or cause.

Accusative
The accusative case (abbreviated acc) indicates the direct object of a verb.

Comitative
The comitative case (abbreviated com), also known as the associative case (abbreviated ass), is a grammatical case that denotes companionship, and is used where English would typically use preposition "with" in the sense of "in company with" or "together with" (other uses of "with," e.g. with the meaning of "using," "by means of" (I cut bread with a knife) would correspond to the instrumental case or related cases).

Dative
The dative case (abbreviated dat, or sometimes d when it is a core argument) indicates the indirect object of a verb.

Delative
The delative case (abbreviated del; from Latin deferre "to bear or bring away or down") in the Hungarian language can originally express the movement from the surface of something (e.g. "off the table"), but it is used in several other meanings (e.g. "about people"), some of them related to the original (e.g. "from the post office").

Distributive
The distributive case (abbreviated distr) is used on nouns for the meanings of per or each.

DistributiveTemporal
The distributive-temporal case specifies when something is done.

Essive
The essive or similaris case (abbreviated ess) carries the meaning of a temporary location or state of being, often equivalent to the English "as a (child)".

Genitive
The genitive case (abbreviated gen; also called the possessive case or second case), which roughly corresponds to English's possessive case and preposition of, indicates the possessor of another noun.

Instrumental
The instrumental case (abbreviated ins or instr; also called the eighth case) indicates an object used in performing an action.

Instructive
In the Finnish language and Estonian language, the instructive case has the basic meaning of "by means of". It is a comparatively rarely used case, though it is found in some commonly used expressions, such as omin silmin -> "with one's own eyes".

Lative
Lative (abbreviated lat) is a case which indicates motion to a location. It corresponds to the English prepositions "to" and "into". The lative case belongs to the group of the general local cases together with the locative and separative case. The term derives from the Latin lat-, the participle stem of ferre, "to bring".

Locative (Maybe Locative)
The locative case (abbreviated loc) indicates a location.

Multiplicative
The multiplicative case is a grammatical case used for marking a number of something ("three times").

Nominative
The nominative case (abbreviated nom) indicates the subject of a finite verb.

Partitive
The partitive case (abbreviated ptv or more ambiguously part) denotes "partialness", "without result", or "without specific identity". It is also used in contexts where a subgroup is selected from a larger group, or with numbers.

Sublative
The term sublative case (abbreviated subl) is used to refer to grammatical cases expressing different situations: In Hungarian, it expresses the destination of the movement, originally to the surface of something (e.g. sit down on the ground, climb the tree), but in other figurative meanings as well (e.g. to university, for two nights), while in Tsez and other Northeast Caucasian languages it denotes a movement towards the bottomsides or the area under an object. The sublative case is used in the Finnish, Tsez and Hungarian languages.

SuperEssive
The Superessive case (abbreviated supe) is a grammatical declension indicating location on top of, or on the surface of something. Its name comes from Latin supersum, superesse: to be over and above.

Translative
The translative case (abbreviated transl) is a grammatical case that indicates a change in state of a noun, with the general sense of "becoming X" or "change to X".

Vocative
The vocative case indicates an addressee.

data Gender

Constructors

Neuter

Masculine

Feminine

Common

data Number

Constructors

Singular

Dual

Trial

Paucal

Plural

Grammatical categories

Case

noCase :: Inflection -> Inflection

hasNoCase :: Inflection -> Bool

ablative :: Inflection -> Inflection

isAblative :: Inflection -> Bool

abessive :: Inflection -> Inflection

isAbessive :: Inflection -> Bool

accusative :: Inflection -> Inflection

isAccusative :: Inflection -> Bool

comitative :: Inflection -> Inflection

isComitative :: Inflection -> Bool

dative :: Inflection -> Inflection

isDative :: Inflection -> Bool

delative :: Inflection -> Inflection

isDelative :: Inflection -> Bool

distributive :: Inflection -> Inflection

isDistributive :: Inflection -> Bool

distributiveTemporal :: Inflection -> Inflection

isDistributiveTemporal :: Inflection -> Bool

essive :: Inflection -> Inflection

isEssive :: Inflection -> Bool

genitive :: Inflection -> Inflection

isGenitive :: Inflection -> Bool

instrumental :: Inflection -> Inflection

isInstrumental :: Inflection -> Bool

instructive :: Inflection -> Inflection

isInstructive :: Inflection -> Bool

lative :: Inflection -> Inflection

isLative :: Inflection -> Bool

locative :: Inflection -> Inflection

isLocative :: Inflection -> Bool

multiplicative :: Inflection -> Inflection

isMultiplicative :: Inflection -> Bool

nominative :: Inflection -> Inflection

isNominative :: Inflection -> Bool

partitive :: Inflection -> Inflection

isPartitive :: Inflection -> Bool

superEssive :: Inflection -> Inflection

isSuperEssive :: Inflection -> Bool

sublative :: Inflection -> Inflection

isSublative :: Inflection -> Bool

translative :: Inflection -> Inflection

isTranslative :: Inflection -> Bool

vocative :: Inflection -> Inflection

isVocative :: Inflection -> Bool

Locative cases

locativeInessive :: Inflection -> Inflection

isLocativeInessive :: Inflection -> Bool

locativeElative :: Inflection -> Inflection

isLocativeElative :: Inflection -> Bool

locativeIllative :: Inflection -> Inflection

isLocativeIllative :: Inflection -> Bool

locativeAdessive :: Inflection -> Inflection

isLocativeAdessive :: Inflection -> Bool

locativeAblative :: Inflection -> Inflection

isLocativeAblative :: Inflection -> Bool

locativeAllative :: Inflection -> Inflection

isLocativeAllative :: Inflection -> Bool

Gender

noGender :: Inflection -> Inflection

hasNoGender :: Inflection -> Bool

neuter :: Inflection -> Inflection

isNeuter :: Inflection -> Bool

masculine :: Inflection -> Inflection

isMasculine :: Inflection -> Bool

feminine :: Inflection -> Inflection

isFeminine :: Inflection -> Bool

common :: Inflection -> Inflection

isCommon :: Inflection -> Bool

Number

noNumber :: Inflection -> Inflection

hasNoNumber :: Inflection -> Bool

singular :: Inflection -> Inflection

isSingular :: Inflection -> Bool

dual :: Inflection -> Inflection

isDual :: Inflection -> Bool

trial :: Inflection -> Inflection

isTrial :: Inflection -> Bool

paucal :: Inflection -> Inflection

isPaucal :: Inflection -> Bool

plural :: Inflection -> Inflection

isPlural :: Inflection -> Bool